refactor(KonnectorAction): Extract InformativeModal
